refactor: extract remote tmux command sending
